Added curl and curl flag option on bash
Added an option for non-Linux users to select between using wget or curl to download files using this script. To use curl, simply include the -v curl flag when running the script. If no flag is specified, wget is used by default. This allows for greater flexibility and ease of use across different platforms.