Use '-nc' instead of '-c'
Excerpt from wget docs:

If a file is downloaded more than once in the same directory, Wget's behavior depends on a few options, including -nc.  In certain cases, the local file will be clobbered, or overwritten, upon repeated download.
In other cases it will be preserve

When running Wget without -N, -nc, -r, or -p, downloading the same file in the same directory will result in the original copy of file being preserved and the second copy being named file.1.  If that file is downloaded yet again, the third copy will be named file.2, and so on.  (This is also the behavior with -nd, even if -r or -p are in effect.)  When -nc is specified, this behavior is suppressed, and Wget will refuse to download newer copies of file.  Therefore, ""no-clobber"" is actually a misnomer in this mode---it's not clobbering that's prevented (as the numeric suffixes were already preventing clobbering), but rather the multiple version saving that's prevented.